[♻️Refactor]: 함수 분리 작업
- 리팩토링 중입니다.
