+++ src/lib/detail/accordion.js
@@ -1,13 +1,14 @@
-import { secretCheck } from './secretCheck';
-import { toggleContent } from './toggleContent';
-
+import { secretCheck } from './secretCheck.js';
+import { toggleContent } from './toggleContent.js';
+import { getNodes } from './getNode.js';
 export function accordion(inquiry, userName, createDay) {
   // 리뷰 섹션
-  const reviewTitle = document.querySelectorAll('.review_title');
-  const reviewContents = document.querySelectorAll('.review_content');
+
+  const reviewTitle = getNodes('.review_title');
+  const reviewContents = getNodes('.review_content');
   // 문의 섹션
-  const inquiryTitle = document.querySelectorAll('.inquiry_title');
-  const inquiryContents = document.querySelectorAll('.inquiry_content');
+  const inquiryTitle = getNodes('.inquiry_title');
+  const inquiryContents = getNodes('.inquiry_content');
 
   secretCheck(inquiryTitle, inquiry, userName, createDay);
   toggleContent(reviewTitle, reviewContents);
