+++ src/lib/detail/updateHeartIcon.js
@@ -1,0 +1,6 @@
+export function updateHeartIcon(iconHeart, productWish) {
+  iconHeart.setAttribute(
+    'href',
+    `/public/icon/_sprite.svg#Heart${productWish ? '-active' : ''}`
+  );
+}
